move field func logic to BaseSource class
