+++ magpylib/_src/obj_classes/class_BaseExcitations.py
@@ -2,20 +2,47 @@
 DOCSTRINGS V4 READY
 """
 from magpylib._src.fields.field_wrap_BH import getBH_level2
-from magpylib._src.utility import format_star_input
+from magpylib._src.input_checks import check_format_input_scalar
 from magpylib._src.input_checks import check_format_input_vector
+from magpylib._src.input_checks import validate_field_func
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
-from magpylib._src.input_checks import check_format_input_scalar
+from magpylib._src.utility import format_star_input
+
 
 class BaseSource(BaseGeo, BaseDisplayRepr):
     """Base class for all types of sources. Provides getB and getH methods for source objects
     and corresponding field function"""
 
-    def __init__(self, position, orientation, style, **kwargs):
-        BaseGeo.__init__(self, position, orientation, style, **kwargs)
+    _field_func = None
+    _editable_field_func = False
+
+    def __init__(self, position, orientation, field_func=None, style=None, **kwargs):
+        if field_func is not None:
+            self.field_func = field_func
+        BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
         BaseDisplayRepr.__init__(self)
 
+    @property
+    def field_func(self):
+        """
+        The function for B- and H-field computation must have the two positional arguments
+        `field` and `observers`. With `field='B'` or `field='H'` the B- or H-field in units
+        of [mT] or [kA/m] must be returned respectively. The `observers` argument must
+        accept numpy ndarray inputs of shape (n,3), in which case the returned fields must
+        be numpy ndarrays of shape (n,3) themselves.
+        """
+        return self._field_func
+
+    @field_func.setter
+    def field_func(self, val):
+        if self._editable_field_func:
+            validate_field_func(val)
+        else:
+            raise AttributeError(
+                "The `field_func` attribute should not be edited for original Magpylib sources."
+            )
+        self._field_func = val
 
     def getB(self, *observers, squeeze=True, pixel_agg=None, output="ndarray"):
         """Compute the B-field in units of [mT] generated by the source.
@@ -161,7 +188,7 @@ class BaseHomMag(BaseSource):
     """provides the magnetization attribute  for homogeneously magnetized magnets"""
 
     def __init__(self, position, orientation, magnetization, style, **kwargs):
-        super().__init__(position, orientation, style, **kwargs)
+        super().__init__(position, orientation, style=style, **kwargs)
         self.magnetization = magnetization
 
     @property
@@ -186,7 +213,7 @@ class BaseCurrent(BaseSource):
     """provides scalar current attribute"""
 
     def __init__(self, position, orientation, current, style, **kwargs):
-        super().__init__(position, orientation, style, **kwargs)
+        super().__init__(position, orientation, style=style, **kwargs)
         self.current = current
 
     @property
